src/requirements.txt
SQLAlchemy==1.3.11
nose==1.3.7
psycopg2==2.8.4
Flask-Bcrypt==0.7.1
Flask-SQLAlchemy===2.4.1
Flask-Migrate==2.5.2
python-dateutil==2.4.2
Flask==1.1.1
apscheduler==3.6.3
python-dotenv==0.12.0
flask-jwt-extended==3.24.1
git+https://github.com/velopaymentsapi/velo-python.git@2.20.29-beta.1